Add constructs directory as a ts path
