--- src/classes/index.ts
@@ -1,3 +1,0 @@
-export { AwsConstruct } from "./AwsConstruct";
-export type { ConstructInterface, StaticConstructInterface } from "./Construct";
-export type { ProviderInterface, StaticProviderInterface } from "./Provider";
