+++ src/constructs/aws/DatabaseDynamoDBSingleTable.ts
@@ -2,8 +2,8 @@ import { Construct as CdkConstruct, CfnOutput, Fn, Stack } from "@aws-cdk/core";
 import { AttributeType, BillingMode, StreamViewType, Table } from "@aws-cdk/aws-dynamodb";
 import { FromSchema } from "json-schema-to-ts";
 import { AwsProvider } from "@lift/providers";
-import { AwsConstruct } from "../classes";
-import { PolicyStatement } from "../CloudFormation";
+import { AwsConstruct } from "@lift/constructs/aws";
+import { PolicyStatement } from "../../CloudFormation";
 
 const DATABASE_DEFINITION = {
     type: "object",
