+++ src/constructs/aws/Storage.ts
@@ -2,8 +2,8 @@ import { BlockPublicAccess, Bucket, BucketEncryption, StorageClass } from "@aws-
 import { Construct as CdkConstruct, CfnOutput, Duration, Fn, Stack } from "@aws-cdk/core";
 import { FromSchema } from "json-schema-to-ts";
 import { AwsProvider } from "@lift/providers";
-import { AwsConstruct } from "../classes";
-import { PolicyStatement } from "../CloudFormation";
+import { AwsConstruct } from "@lift/constructs/aws";
+import { PolicyStatement } from "../../CloudFormation";
 
 const STORAGE_DEFINITION = {
     type: "object",
