+++ src/plugin.ts
@@ -6,7 +6,8 @@ import { readFileSync } from "fs";
 import { dump } from "js-yaml";
 import { DefaultTokenResolver, Lazy, StringConcat, Tokenization } from "@aws-cdk/core";
 import { FromSchema } from "json-schema-to-ts";
-import { AwsProvider, StripeProvider } from "@lift/providers";
+import { AwsProvider, ProviderInterface, StaticProviderInterface, StripeProvider } from "@lift/providers";
+import { ConstructInterface, StaticConstructInterface } from "@lift/constructs";
 import type {
     CommandsDefinition,
     DeprecatedVariableResolver,
@@ -14,7 +15,6 @@ import type {
     Serverless,
     VariableResolver,
 } from "./types/serverless";
-import { ConstructInterface, ProviderInterface, StaticConstructInterface, StaticProviderInterface } from "./classes";
 import { log } from "./utils/logger";
 import ServerlessError from "./utils/error";
