+++ src/providers/AwsProvider.ts
@@ -1,19 +1,13 @@
 import { App, CfnOutput, Stack } from "@aws-cdk/core";
 import { get, merge } from "lodash";
 import { AwsCfInstruction, AwsLambdaVpcConfig } from "@serverless/typescript";
+import { ProviderInterface } from "@lift/providers";
+import { ConstructInterface, StaticConstructInterface } from "@lift/constructs";
+import { DatabaseDynamoDBSingleTable, Queue, StaticWebsite, Storage, Vpc, Webhook } from "@lift/constructs/aws";
 import { getStackOutput } from "../CloudFormation";
 import { CloudformationTemplate, Provider as LegacyAwsProvider, Serverless } from "../types/serverless";
 import { awsRequest } from "../classes/aws";
-import { ConstructInterface } from "../classes";
-import { StaticConstructInterface } from "../classes/Construct";
 import ServerlessError from "../utils/error";
-import { Storage } from "../constructs/Storage";
-import { Queue } from "../constructs/Queue";
-import { Vpc } from "../constructs/Vpc";
-import { Webhook } from "../constructs/Webhook";
-import { StaticWebsite } from "../constructs/StaticWebsite";
-import { DatabaseDynamoDBSingleTable } from "../constructs/DatabaseDynamoDBSingleTable";
-import { ProviderInterface } from "../classes/Provider";
 
 export class AwsProvider implements ProviderInterface {
     private static readonly constructClasses: Record<string, StaticConstructInterface> = {};
